some updates with update week cron job
