+++ lib/actions/session.server.actions.ts
@@ -49,36 +49,7 @@ async function isSessioninPastWeek(enrollmentId: string, midWeek: Date) {
   return Object.keys(data).length > 0;
 }
 
-export async function getSessionKeys(data?: Session[]) {
-  const supabase = await createClient();
-  const sessionKeys: Set<string> = new Set();
 
-  if (!data) {
-    const { data, error } = await supabase
-      .from(Table.Sessions)
-      .select("student_id, tutor_id, date");
-
-    if (error) {
-      console.error("Error fetching sessions:", error);
-      throw error;
-    }
-  }
-
-  if (!data) return sessionKeys;
-
-  data.forEach((session) => {
-    if (session.date) {
-      const sessionDate = new Date(session.date);
-      const key = `${session.student?.id}-${session.tutor?.id}-${format(
-        sessionDate,
-        "yyyy-MM-dd-HH:mm"
-      )}`;
-      sessionKeys.add(key);
-    }
-  });
-
-  return sessionKeys;
-}
 
 /**
  * Add sessions for enrollments within the specified week range
@@ -88,211 +59,6 @@ export async function getSessionKeys(data?: Session[]) {
  * @param sessions - Existing sessions to avoid duplicates
  * @returns Newly created sessions
  */
-export async function addSessionsServer(
-  weekStartString: string,
-  weekEndString: string,
-  enrollments: Enrollment[],
-  sessions: Session[]
-) {
-  const supabase = await createClient();
-
-  try {
-    const weekStart: Date = fromZonedTime(
-      parseISO(weekStartString),
-      "America/New_York"
-    );
-    const weekEnd: Date = fromZonedTime(
-      parseISO(weekEndString),
-      "America/New_York"
-    );
-
-    const now: string = new Date().toISOString();
-
-    //Set created to avoid duplicates
-    const scheduledSessions: Set<string> = await getSessionKeys(sessions);
-    // Prepare bulk insert data
-    const sessionsToCreate: any[] = [];
-
-    // Process all enrollments
-    for (const enrollment of enrollments) {
-      const {
-        id,
-        student,
-        tutor,
-        availability,
-        meetingId,
-        summary,
-        startDate,
-        duration,
-        frequency,
-      } = enrollment;
-
-      // if (frequency === "biweekly") {
-      //   if (await isSessioninPastWeek(id, addDays(weekStart, 3))) continue;
-      // }
-
-      const startDate_asDate = new Date(startDate); //UTC
-
-      //Check if paused over the summer
-      if (enrollment.paused) {
-        continue;
-      }
-
-      // Skip invalid enrollments
-      if (!student?.id || !tutor?.id || !availability?.length) {
-        continue;
-      }
-
-      // Process each availability slot
-      let { day, startTime, endTime } = availability[0];
-
-      // Skip invalid time formats
-      if (
-        !startTime ||
-        !endTime
-        // startTime.includes("-") ||
-        // endTime.includes("-")
-      ) {
-        console.error(`Invalid time format in availability:`, availability[0]);
-        continue;
-      }
-
-      // Find matching day in the week range
-      let currentDate = new Date(weekStart);
-      const dayLower = day.toLowerCase();
-
-      while (currentDate <= weekEnd) {
-        const currentDay = format(currentDate, "EEEE").toLowerCase();
-
-        // Skip days that don't match
-        if (currentDay !== dayLower) {
-          currentDate = addDays(currentDate, 1);
-          continue;
-        }
-
-        //Add Seven Days if CurrentDate is last week (Acts as a Modulus to ensure updating current week only)
-        if (currentDate < parseISO(weekStartString)) {
-          currentDate = addDays(currentDate, 7);
-        }
-
-        //Remove Seven Days if CurrentDate is next week (Acts as a Modulus to ensure updating current week only)
-        if (currentDate > parseISO(weekEndString)) {
-          currentDate = addDays(currentDate, -7);
-        }
-
-        try {
-          // Parse times correctly
-          const [startHour, startMinute] = startTime.split(":").map(Number);
-          const [endHour, endMinute] = endTime.split(":").map(Number);
-
-          if (
-            isNaN(startHour) ||
-            isNaN(startMinute) ||
-            isNaN(endHour) ||
-            isNaN(endMinute)
-          ) {
-            throw new Error(
-              `Invalid time format: start=${startTime}, end=${endTime}`
-            );
-          }
-
-          // Create session date with correct time
-          // * SetHours and SetMinutes are dependent on local timezone
-
-          const dateString = `${format(currentDate, "yyyy-MM-dd")}T${startTime}:00`;
-          const sessionStartTime = fromZonedTime(
-            dateString,
-            "America/New_York"
-          ); // Automatically handles DST
-
-          if (sessionStartTime < startDate_asDate) {
-            throw new Error("Session occurs before start date");
-          }
-
-          // Check for duplicate session
-          const sessionKey = `${student.id}-${tutor.id}-${format(
-            sessionStartTime,
-            "yyyy-MM-dd-HH:mm"
-          )}`;
-
-          if (!scheduledSessions.has(sessionKey)) {
-            // Add to batch insert
-            sessionsToCreate.push({
-              enrollment_id: id,
-              date: sessionStartTime.toISOString(),
-              student_id: student.id,
-              tutor_id: tutor.id,
-              status: "Active",
-              summary: summary || "",
-              meeting_id: meetingId || null,
-              duration: duration,
-            });
-
-            // Track this session to avoid duplicates
-            scheduledSessions.add(sessionKey);
-          } ////
-        } catch (err) {
-          console.error(
-            "Error processing time for %s %s-%s:",
-            day,
-            startTime,
-            endTime,
-            err
-          );
-        }
-
-        // Move to next day
-        currentDate = addDays(currentDate, 1);
-      }
-    }
-
-    // Perform batch insert if we have sessions to create
-    if (sessionsToCreate.length > 0) {
-      const { data, error } = await supabase
-        .from(Table.Sessions)
-        .insert(sessionsToCreate).select(`
-          *,
-          student:Profiles!student_id(*),
-          tutor:Profiles!tutor_id(*),
-          meeting:Meetings!meeting_id(*)
-          `);
-
-      if (error) throw error;
-
-      if (data) {
-        // Transform returned data to Session objects
-        const sessions: Session[] = data.map((session: any) => ({
-          id: session.id,
-          enrollmentId: session.enrollment_id,
-          createdAt: session.created_at,
-          environment: session.environment,
-          date: session.date,
-          summary: session.summary,
-          meeting: session.meeting,
-          student: session.student,
-          tutor: session.tutor,
-          status: session.status,
-          session_exit_form: session.session_exit_form || null,
-          isQuestionOrConcern: session.isQuestionOrConcern,
-          isFirstSession: session.isFirstSession,
-          duration: session.duration,
-        }));
-
-        // // if (!sessions) return;
-
-        // scheduleMultipleSessionReminders(sessions!);
-
-        // //Schedule emails
-        return sessions;
-      }
-    }
-
-    return [];
-  } catch (error) {
-    console.error("Error creating sessions:", error);
-    throw error;
-  }
-}
 
 /**
  * Normalize meeting ID by removing spaces for comparison
